Update findRecord.sh
Now, findRecord expects a parameter and will find any partial matches to any of the fields in a database entry.

For example, I added a Michael Baum and a Michael Smith. If you run ./findRecord.sh Michael, it will find both of them.

It returns (exports) a variable to the menu named $found_results. This contains the lines that were found, and they are separated by a ;